# Conflicts:
#	D3D_DEMO_231011/D3D_DEMO_231011/Assets.cpp
#	D3D_DEMO_231011/D3D_DEMO_231011/Assets.h
#	D3D_DEMO_231011/D3D_DEMO_231011/Game.cpp
#	D3D_DEMO_231011/D3D_DEMO_231011/Game.h
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/Assets.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/ColliderManager.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/D3D_DEMO_231011.ilk
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/D3D_DEMO_231011.log
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/D3D_DEMO_231011.tlog/CL.command.1.tlog
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/D3D_DEMO_231011.tlog/CL.read.1.tlog
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/D3D_DEMO_231011.tlog/CL.write.1.tlog
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/DebugManager.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/Effect.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/Game.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/GameObject.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/RailManager.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/Result.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/ResultAnimation.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/ResultProcess.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/SceneManager.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/Sprite.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/WinMain.sbr
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/vc143.idb
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/vc143.pdb
#	D3D_DEMO_231011/D3D_DEMO_231011/x64/Debug/xa2.sbr
#	D3D_DEMO_231011/x64/Debug/D3D_DEMO_231011.bsc
#	D3D_DEMO_231011/x64/Debug/D3D_DEMO_231011.pdb

+++ D3D_DEMO_231011/D3D_DEMO_231011/Collider.cpp
@@ -27,8 +27,8 @@ void Collider::Update(DirectX::XMFLOAT3 center, DirectX::XMFLOAT3 rotation, Dire
 Collider_Dir Collider::GetCollisionDir(Collider* collider)
 {
 	Collider_Dir dir = { INI_STATE,INI_STATE };
-	int x_Result = this->m_center.x - collider->m_center.x;
-	int y_Result = this->m_center.y - collider->m_center.y;
+	int x_Result = (int)(this->m_center.x - collider->m_center.x);
+	int y_Result = (int)(this->m_center.y - collider->m_center.y);
 
 	//TargetObject は本体の左
 	if (x_Result > 0) {
